ml-worker/requirements.txt
cloudinary>=1.40.0
insightface>=0.7.3
numpy>=1.24.0
onnxruntime>=1.16.0
opencv-python-headless>=4.9.0
pika>=1.3.2
python-dotenv>=1.0.1
requests>=2.31.0
psycopg2-binary>=2.9.9
